(rigobersio)edit:express cors y res.cookie(ajustes para token al 99%)
