+++ server/src/App.js
@@ -9,6 +9,7 @@ import taskRoutes from "./routes/tasks.routes.js";
 const app = express();
 app.use(cors({
   origin: 'http://localhost:5173',
+  credentials: true,
 }));
 
 //app.use(cors());
